requirements.txt
asgiref==3.3.4
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0.5
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.11.2
django-g-recaptcha==0.2.4
django-intl-tel-input==0.3.1
django-location-field==2.1.0
django-mapbox-location-field==1.7.3
django-octicons==1.0.2
django-phonenumber-field==5.1.0
djongo==1.3.4
dnspython==1.16.0
funcsigs==1.0.2
idna==2.10
mock==4.0.3
oauthlib==3.1.0
pbr==5.6.0
phonenumbers==8.12.21
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
urllib3==1.26.4
